docker, vector, README: Sink logs to Elasticsearch
Update the docker-compose to include images for Elasticsearch and
Kibana. This is done to now sink the output logs to an Elasticsearch
database and view the logs via a web interface - Kibana.
Comment out the file sinking from the vector.toml file.
Update the README instructions on how to run, view the logs and
clean up the docker images.

Signed-off-by: Joshua Van Straaten <joshuavanstraaten100@gmail.com>